Split field tests up, write some m2m tests
