--- tortoise/tests/test_fields.py
@@ -1,627 +1,0 @@
-from datetime import date, datetime, timedelta
-from decimal import Decimal
-from time import sleep
-
-from tortoise import fields
-from tortoise.contrib import test
-from tortoise.exceptions import (ConfigurationError, IntegrityError, NoValuesFetched,
-                                 OperationalError)
-from tortoise.tests import testmodels
-
-
-class TestRequired(test.SimpleTestCase):
-
-    async def test_required_by_default(self):
-        self.assertTrue(fields.Field(str).required)
-
-    async def test_if_generated_then_not_required(self):
-        self.assertFalse(fields.Field(str, generated=True).required)
-
-    async def test_if_null_then_not_required(self):
-        self.assertFalse(fields.Field(str, null=True).required)
-
-    async def test_if_has_non_null_default_then_not_required(self):
-        self.assertFalse(fields.Field(str, default='').required)
-
-    async def test_if_null_default_then_required(self):
-        self.assertTrue(fields.Field(str, default=None).required)
-
-
-class TestIntFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.IntFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.IntFields.create(intnum=2147483647)
-        obj = await testmodels.IntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, 2147483647)
-        self.assertEqual(obj.intnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.IntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_min(self):
-        obj0 = await testmodels.IntFields.create(intnum=-2147483648)
-        obj = await testmodels.IntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, -2147483648)
-        self.assertEqual(obj.intnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.IntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_cast(self):
-        obj0 = await testmodels.IntFields.create(intnum='3')
-        obj = await testmodels.IntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, 3)
-
-    async def test_values(self):
-        obj0 = await testmodels.IntFields.create(intnum=1)
-        values = await testmodels.IntFields.get(id=obj0.id).values('intnum')
-        self.assertEqual(values[0]['intnum'], 1)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.IntFields.create(intnum=1)
-        values = await testmodels.IntFields.get(id=obj0.id).values_list('intnum', flat=True)
-        self.assertEqual(values[0], 1)
-
-
-class TestSmallIntFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.SmallIntFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.SmallIntFields.create(smallintnum=32767)
-        obj = await testmodels.SmallIntFields.get(id=obj0.id)
-        self.assertEqual(obj.smallintnum, 32767)
-        self.assertEqual(obj.smallintnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.SmallIntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_min(self):
-        obj0 = await testmodels.SmallIntFields.create(smallintnum=-32768)
-        obj = await testmodels.SmallIntFields.get(id=obj0.id)
-        self.assertEqual(obj.smallintnum, -32768)
-        self.assertEqual(obj.smallintnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.SmallIntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.SmallIntFields.create(smallintnum=2)
-        values = await testmodels.SmallIntFields.get(id=obj0.id).values('smallintnum')
-        self.assertEqual(values[0]['smallintnum'], 2)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.SmallIntFields.create(smallintnum=2)
-        values = await testmodels.SmallIntFields.get(
-            id=obj0.id).values_list('smallintnum', flat=True)
-        self.assertEqual(values[0], 2)
-
-
-class TestBigIntFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.BigIntFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.BigIntFields.create(intnum=9223372036854775807)
-        obj = await testmodels.BigIntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, 9223372036854775807)
-        self.assertEqual(obj.intnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.BigIntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_min(self):
-        obj0 = await testmodels.BigIntFields.create(intnum=-9223372036854775808)
-        obj = await testmodels.BigIntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, -9223372036854775808)
-        self.assertEqual(obj.intnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.BigIntFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_cast(self):
-        obj0 = await testmodels.BigIntFields.create(intnum='3')
-        obj = await testmodels.BigIntFields.get(id=obj0.id)
-        self.assertEqual(obj.intnum, 3)
-
-    async def test_values(self):
-        obj0 = await testmodels.BigIntFields.create(intnum=1)
-        values = await testmodels.BigIntFields.get(id=obj0.id).values('intnum')
-        self.assertEqual(values[0]['intnum'], 1)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.BigIntFields.create(intnum=1)
-        values = await testmodels.BigIntFields.get(id=obj0.id).values_list('intnum', flat=True)
-        self.assertEqual(values[0], 1)
-
-
-class TestCharFields(test.TestCase):
-    def test_max_length_missing(self):
-        with self.assertRaisesRegex(TypeError,
-                                    "missing 1 required positional argument: 'max_length'"):
-            fields.CharField()
-
-    def test_max_length_bad(self):
-        with self.assertRaisesRegex(ConfigurationError, "'max_length' must be >= 1"):
-            fields.CharField(max_length=0)
-
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.CharFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.CharFields.create(char='moo')
-        obj = await testmodels.CharFields.get(id=obj0.id)
-        self.assertEqual(obj.char, 'moo')
-        self.assertEqual(obj.char_null, None)
-        await obj.save()
-        obj2 = await testmodels.CharFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_cast(self):
-        obj0 = await testmodels.CharFields.create(char=33)
-        obj = await testmodels.CharFields.get(id=obj0.id)
-        self.assertEqual(obj.char, '33')
-
-    async def test_values(self):
-        obj0 = await testmodels.CharFields.create(char='moo')
-        values = await testmodels.CharFields.get(id=obj0.id).values('char')
-        self.assertEqual(values[0]['char'], 'moo')
-
-    async def test_values_list(self):
-        obj0 = await testmodels.CharFields.create(char='moo')
-        values = await testmodels.CharFields.get(id=obj0.id).values_list('char', flat=True)
-        self.assertEqual(values[0], 'moo')
-
-
-class TestTextFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.TextFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.TextFields.create(text='baa')
-        obj = await testmodels.TextFields.get(id=obj0.id)
-        self.assertEqual(obj.text, 'baa')
-        self.assertEqual(obj.text_null, None)
-        await obj.save()
-        obj2 = await testmodels.TextFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.TextFields.create(text='baa')
-        values = await testmodels.TextFields.get(id=obj0.id).values('text')
-        self.assertEqual(values[0]['text'], 'baa')
-
-    async def test_values_list(self):
-        obj0 = await testmodels.TextFields.create(text='baa')
-        values = await testmodels.TextFields.get(id=obj0.id).values_list('text', flat=True)
-        self.assertEqual(values[0], 'baa')
-
-
-class TestBooleanFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.BooleanFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.BooleanFields.create(boolean=True)
-        obj = await testmodels.BooleanFields.get(id=obj0.id)
-        self.assertEqual(obj.boolean, True)
-        self.assertEqual(obj.boolean_null, None)
-        await obj.save()
-        obj2 = await testmodels.BooleanFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.BooleanFields.create(boolean=True)
-        values = await testmodels.BooleanFields.get(id=obj0.id).values('boolean')
-        self.assertEqual(values[0]['boolean'], True)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.BooleanFields.create(boolean=True)
-        values = await testmodels.BooleanFields.get(id=obj0.id).values_list('boolean', flat=True)
-        self.assertEqual(values[0], True)
-
-
-class TestDecimalFields(test.TestCase):
-    def test_max_digits_empty(self):
-        with self.assertRaisesRegex(TypeError,
-                                    "missing 2 required positional arguments: 'max_digits' and"
-                                    " 'decimal_places'"):
-            fields.DecimalField()
-
-    def test_decimal_places_empty(self):
-        with self.assertRaisesRegex(TypeError,
-                                    "missing 1 required positional argument: 'decimal_places'"):
-            fields.DecimalField(max_digits=1)
-
-    def test_max_fields_bad(self):
-        with self.assertRaisesRegex(ConfigurationError, "'max_digits' must be >= 1"):
-            fields.DecimalField(max_digits=0, decimal_places=2)
-
-    def test_decimal_places_bad(self):
-        with self.assertRaisesRegex(ConfigurationError, "'decimal_places' must be >= 0"):
-            fields.DecimalField(max_digits=2, decimal_places=-1)
-
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.DecimalFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.DecimalFields.create(decimal=Decimal('1.23456'), decimal_nodec=18.7)
-        obj = await testmodels.DecimalFields.get(id=obj0.id)
-        self.assertEqual(obj.decimal, Decimal('1.2346'))
-        self.assertEqual(obj.decimal_nodec, 19)
-        self.assertEqual(obj.decimal_null, None)
-        await obj.save()
-        obj2 = await testmodels.DecimalFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.DecimalFields.create(decimal=Decimal('1.23456'), decimal_nodec=18.7)
-        values = await testmodels.DecimalFields.get(id=obj0.id).values('decimal', 'decimal_nodec')
-        self.assertEqual(values[0]['decimal'], Decimal('1.2346'))
-        self.assertEqual(values[0]['decimal_nodec'], 19)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.DecimalFields.create(decimal=Decimal('1.23456'), decimal_nodec=18.7)
-        values = await testmodels.DecimalFields.get(
-            id=obj0.id).values_list('decimal', 'decimal_nodec')
-        self.assertEqual(list(values[0]), [Decimal('1.2346'), 19])
-
-
-class TestDatetimeFields(test.TestCase):
-    def test_both_auto_bad(self):
-        with self.assertRaisesRegex(ConfigurationError,
-                                    "You can choose only 'auto_now' or 'auto_now_add'"):
-            fields.DatetimeField(auto_now=True, auto_now_add=True)
-
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.DatetimeFields.create()
-
-    async def test_create(self):
-        now = datetime.utcnow()
-        obj0 = await testmodels.DatetimeFields.create(datetime=now)
-        obj = await testmodels.DatetimeFields.get(id=obj0.id)
-        self.assertEqual(obj.datetime, now)
-        self.assertEqual(obj.datetime_null, None)
-        self.assertLess(obj.datetime_auto - now, timedelta(microseconds=10000))
-        self.assertLess(obj.datetime_add - now, timedelta(microseconds=10000))
-        datetime_auto = obj.datetime_auto
-        sleep(0.011)
-        await obj.save()
-        obj2 = await testmodels.DatetimeFields.get(id=obj.id)
-        self.assertEqual(obj2.datetime, now)
-        self.assertEqual(obj2.datetime_null, None)
-        self.assertEqual(obj2.datetime_auto, obj.datetime_auto)
-        self.assertNotEqual(obj2.datetime_auto, datetime_auto)
-        self.assertGreater(obj2.datetime_auto - now, timedelta(microseconds=10000))
-        self.assertLess(obj2.datetime_auto - now, timedelta(seconds=1))
-        self.assertEqual(obj2.datetime_add, obj.datetime_add)
-
-    async def test_cast(self):
-        now = datetime.utcnow()
-        obj0 = await testmodels.DatetimeFields.create(datetime=now.isoformat())
-        obj = await testmodels.DatetimeFields.get(id=obj0.id)
-        self.assertEqual(obj.datetime, now)
-
-    async def test_values(self):
-        now = datetime.utcnow()
-        obj0 = await testmodels.DatetimeFields.create(datetime=now)
-        values = await testmodels.DatetimeFields.get(id=obj0.id).values('datetime')
-        self.assertEqual(values[0]['datetime'], now)
-
-    async def test_values_list(self):
-        now = datetime.utcnow()
-        obj0 = await testmodels.DatetimeFields.create(datetime=now)
-        values = await testmodels.DatetimeFields.get(id=obj0.id).values_list('datetime', flat=True)
-        self.assertEqual(values[0], now)
-
-
-class TestDateFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.DateFields.create()
-
-    async def test_create(self):
-        today = date.today()
-        obj0 = await testmodels.DateFields.create(date=today)
-        obj = await testmodels.DateFields.get(id=obj0.id)
-        self.assertEqual(obj.date, today)
-        self.assertEqual(obj.date_null, None)
-        await obj.save()
-        obj2 = await testmodels.DateFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_cast(self):
-        today = date.today()
-        obj0 = await testmodels.DateFields.create(date=today.isoformat())
-        obj = await testmodels.DateFields.get(id=obj0.id)
-        self.assertEqual(obj.date, today)
-
-    async def test_values(self):
-        today = date.today()
-        obj0 = await testmodels.DateFields.create(date=today)
-        values = await testmodels.DateFields.get(id=obj0.id).values('date')
-        self.assertEqual(values[0]['date'], today)
-
-    async def test_values_list(self):
-        today = date.today()
-        obj0 = await testmodels.DateFields.create(date=today)
-        values = await testmodels.DateFields.get(id=obj0.id).values_list('date', flat=True)
-        self.assertEqual(values[0], today)
-
-
-class TestTimeDeltaFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.TimeDeltaFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.TimeDeltaFields.create(
-            timedelta=timedelta(days=35, seconds=8, microseconds=1))
-        obj = await testmodels.TimeDeltaFields.get(id=obj0.id)
-        self.assertEqual(obj.timedelta, timedelta(days=35, seconds=8, microseconds=1))
-        self.assertEqual(obj.timedelta_null, None)
-        await obj.save()
-        obj2 = await testmodels.TimeDeltaFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.TimeDeltaFields.create(
-            timedelta=timedelta(days=35, seconds=8, microseconds=1))
-        values = await testmodels.TimeDeltaFields.get(id=obj0.id).values('timedelta')
-        self.assertEqual(values[0]['timedelta'], timedelta(days=35, seconds=8, microseconds=1))
-
-    async def test_values_list(self):
-        obj0 = await testmodels.TimeDeltaFields.create(
-            timedelta=timedelta(days=35, seconds=8, microseconds=1))
-        values = await testmodels.TimeDeltaFields.get(id=obj0.id).values_list(
-            'timedelta', flat=True)
-        self.assertEqual(values[0], timedelta(days=35, seconds=8, microseconds=1))
-
-
-class TestFloatFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.FloatFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.FloatFields.create(floatnum=1.23)
-        obj = await testmodels.FloatFields.get(id=obj0.id)
-        self.assertEqual(obj.floatnum, 1.23)
-        self.assertNotEqual(Decimal(obj.floatnum), Decimal('1.23'))
-        self.assertEqual(obj.floatnum_null, None)
-        await obj.save()
-        obj2 = await testmodels.FloatFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_cast_int(self):
-        obj0 = await testmodels.FloatFields.create(floatnum=123)
-        obj = await testmodels.FloatFields.get(id=obj0.id)
-        self.assertEqual(obj.floatnum, 123)
-
-    async def test_cast_decimal(self):
-        obj0 = await testmodels.FloatFields.create(floatnum=Decimal('1.23'))
-        obj = await testmodels.FloatFields.get(id=obj0.id)
-        self.assertEqual(obj.floatnum, 1.23)
-
-    async def test_values(self):
-        obj0 = await testmodels.FloatFields.create(floatnum=1.23)
-        values = await testmodels.FloatFields.filter(id=obj0.id).values('floatnum')
-        self.assertEqual(values[0]['floatnum'], 1.23)
-
-    async def test_values_list(self):
-        obj0 = await testmodels.FloatFields.create(floatnum=1.23)
-        values = await testmodels.FloatFields.filter(id=obj0.id).values_list('floatnum')
-        self.assertEqual(list(values[0]), [1.23])
-
-
-class TestJSONFields(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.JSONFields.create()
-
-    async def test_create(self):
-        obj0 = await testmodels.JSONFields.create(data={'some': ['text', 3]})
-        obj = await testmodels.JSONFields.get(id=obj0.id)
-        self.assertEqual(obj.data, {'some': ['text', 3]})
-        self.assertEqual(obj.data_null, None)
-        await obj.save()
-        obj2 = await testmodels.JSONFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_list(self):
-        obj0 = await testmodels.JSONFields.create(data=['text', 3])
-        obj = await testmodels.JSONFields.get(id=obj0.id)
-        self.assertEqual(obj.data, ['text', 3])
-        self.assertEqual(obj.data_null, None)
-        await obj.save()
-        obj2 = await testmodels.JSONFields.get(id=obj.id)
-        self.assertEqual(obj, obj2)
-
-    async def test_values(self):
-        obj0 = await testmodels.JSONFields.create(data={'some': ['text', 3]})
-        values = await testmodels.JSONFields.filter(id=obj0.id).values('data')
-        self.assertEqual(values[0]['data'], {'some': ['text', 3]})
-
-    async def test_values_list(self):
-        obj0 = await testmodels.JSONFields.create(data={'some': ['text', 3]})
-        values = await testmodels.JSONFields.filter(id=obj0.id).values_list('data', flat=True)
-        self.assertEqual(values[0], {'some': ['text', 3]})
-
-
-class TestForeignKeyField(test.TestCase):
-    async def test_empty(self):
-        with self.assertRaises(IntegrityError):
-            await testmodels.MinRelation.create()
-
-    async def test_minimal__create_by_id(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.MinRelation.create(tournament_id=tour.id)
-        self.assertEqual(rel.tournament_id, tour.id)
-        self.assertEqual((await tour.minrelations.all())[0], rel)
-
-    @test.expectedFailure
-    async def test_minimal__create_by_name(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.MinRelation.create(tournament=tour)
-        self.assertEqual(rel.tournament, tour)
-        self.assertEqual((await tour.minrelations.all())[0], rel)
-
-    async def test_event__create_by_id(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.Event.create(name='Event1', tournament_id=tour.id)
-        self.assertEqual(rel.tournament_id, tour.id)
-        self.assertEqual((await tour.events.all())[0], rel)
-
-    @test.expectedFailure
-    async def test_event__create_by_name(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.Event.create(name='Event1', tournament=tour)
-        self.assertEqual(rel.tournament, tour)
-        self.assertEqual((await tour.events.all())[0], rel)
-
-    async def test_minimal__uninstantiated_create(self):
-        tour = testmodels.Tournament(name='Team1')
-        with self.assertRaisesRegex(OperationalError, 'You should first call .save()'):
-            await testmodels.MinRelation.create(tournament=tour)
-
-    async def test_minimal__uninstantiated_iterate(self):
-        tour = testmodels.Tournament(name='Team1')
-        with self.assertRaisesRegex(OperationalError,
-                                    "This objects hasn't been instanced, call .save()"):
-            async for _ in tour.minrelations:
-                pass
-
-    async def test_minimal__uninstantiated_await(self):
-        tour = testmodels.Tournament(name='Team1')
-        with self.assertRaisesRegex(OperationalError,
-                                    "This objects hasn't been instanced, call .save()"):
-            await tour.minrelations
-
-    async def test_minimal__unfetched_contains(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        with self.assertRaisesRegex(NoValuesFetched,
-                                    "No values were fetched for this relation,"
-                                    " first use .fetch_related()"):
-            'a' in tour.minrelations  # pylint: disable=W0104
-
-    async def test_minimal__unfetched_iter(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        with self.assertRaisesRegex(NoValuesFetched,
-                                    "No values were fetched for this relation,"
-                                    " first use .fetch_related()"):
-            for _ in tour.minrelations:
-                pass
-
-    async def test_minimal__unfetched_len(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        with self.assertRaisesRegex(NoValuesFetched,
-                                    "No values were fetched for this relation,"
-                                    " first use .fetch_related()"):
-            len(tour.minrelations)
-
-    async def test_minimal__unfetched_bool(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        with self.assertRaisesRegex(NoValuesFetched,
-                                    "No values were fetched for this relation,"
-                                    " first use .fetch_related()"):
-            bool(tour.minrelations)
-
-    async def test_minimal__unfetched_getitem(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        with self.assertRaisesRegex(NoValuesFetched,
-                                    "No values were fetched for this relation,"
-                                    " first use .fetch_related()"):
-            tour.minrelations[0]  # pylint: disable=W0104
-
-    async def test_minimal__instantiated_create(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        await testmodels.MinRelation.create(tournament=tour)
-
-    async def test_minimal__instantiated_iterate(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        async for _ in tour.minrelations:
-            pass
-
-    async def test_minimal__instantiated_await(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        await tour.minrelations
-
-    async def test_minimal__fetched_contains(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.MinRelation.create(tournament=tour)
-        await tour.fetch_related('minrelations')
-        self.assertTrue(rel in tour.minrelations)
-
-    async def test_minimal__fetched_iter(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.MinRelation.create(tournament=tour)
-        await tour.fetch_related('minrelations')
-        self.assertEqual([obj for obj in tour.minrelations], [rel])
-
-    async def test_minimal__fetched_len(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        await testmodels.MinRelation.create(tournament=tour)
-        await tour.fetch_related('minrelations')
-        self.assertEqual(len(tour.minrelations), 1)
-
-    async def test_minimal__fetched_bool(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        await tour.fetch_related('minrelations')
-        self.assertFalse(bool(tour.minrelations))
-        await testmodels.MinRelation.create(tournament=tour)
-        await tour.fetch_related('minrelations')
-        self.assertTrue(bool(tour.minrelations))
-
-    async def test_minimal__fetched_getitem(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        rel = await testmodels.MinRelation.create(tournament=tour)
-        await tour.fetch_related('minrelations')
-        self.assertEqual(tour.minrelations[0], rel)
-
-        with self.assertRaises(IndexError):
-            tour.minrelations[1]  # pylint: disable=W0104
-
-    async def test_event__filter(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        event1 = await testmodels.Event.create(name='Event1', tournament=tour)
-        event2 = await testmodels.Event.create(name='Event2', tournament=tour)
-        self.assertEqual(await tour.events.filter(name='Event1'), [event1])
-        self.assertEqual(await tour.events.filter(name='Event2'), [event2])
-        self.assertEqual(await tour.events.filter(name='Event3'), [])
-
-    async def test_event__all(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        event1 = await testmodels.Event.create(name='Event1', tournament=tour)
-        event2 = await testmodels.Event.create(name='Event2', tournament=tour)
-        self.assertSetEqual(set(await tour.events.all()), {event1, event2})
-
-    async def test_event__order_by(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        event1 = await testmodels.Event.create(name='Event1', tournament=tour)
-        event2 = await testmodels.Event.create(name='Event2', tournament=tour)
-        self.assertEqual(await tour.events.all().order_by('-name'), [event2, event1])
-        self.assertEqual(await tour.events.all().order_by('name'), [event1, event2])
-
-    async def test_event__limit(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        event1 = await testmodels.Event.create(name='Event1', tournament=tour)
-        event2 = await testmodels.Event.create(name='Event2', tournament=tour)
-        await testmodels.Event.create(name='Event3', tournament=tour)
-        self.assertEqual(await tour.events.all().order_by('name').limit(2), [event1, event2])
-
-    async def test_event__offset(self):
-        tour = await testmodels.Tournament.create(name='Team1')
-        await testmodels.Event.create(name='Event1', tournament=tour)
-        event2 = await testmodels.Event.create(name='Event2', tournament=tour)
-        event3 = await testmodels.Event.create(name='Event3', tournament=tour)
-        self.assertEqual(await tour.events.all().order_by('name').offset(1), [event2, event3])
